javascript rounding fix and updates to check return

+++ app/assets/javascripts/transform-paste.js
@@ -22,7 +22,7 @@
                                 if (copyValue) {
                                     var cellData = copyValue.split('\t');
                                     var cellValue = cellData[tableColumnCount - 1];
-                                    cellValue = parseInt(cellValue);
+                                    cellValue = parseFloat(cellValue);
                                     var input = tableRows[tableRowCount].cells[tableColumnCount].querySelector('input:first-of-type');
                                     if (!isNaN(cellValue)) {
                                         var convertedValue = parseFloat(cellValue).toFixed(3);
